Details and Filters Sidepanel
Added and stylized a side panel with two accordion views.  The first view gives details about a selected node or edge.  If a node is selected, the user can select the degrees of separation to view.

The second view is filter toggle options.  The javascript populates the filterlist based on what groups are represented in the nodes.  When a filter is toggled unchecked, the nodes associated with that group are hidden, along with their attached edges.